tests: split fixtures into TestBase class
